apps/chatbot/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
httpx==0.27.2
python-dotenv==1.0.1
qdrant-client==1.9.2
sentence-transformers==2.7.0
numpy==1.26.4
PyYAML==6.0.2

